requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiohttp-retry==2.9.1
aiosignal==1.3.2
amqp==5.3.1
asgiref==3.8.1
attrs==25.3.0
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.5
Django==5.2
django-celery-beat==2.8.0
django-cors-headers==4.7.0
django-environ==0.12.0
django-filter==25.1
django-phonenumber-field==8.1.0
django-timezone-field==7.1
djangorestframework==3.16.0
drf-yasg==1.21.10
frozenlist==1.5.0
idna==3.10
inflection==0.5.1
kombu==5.5.3
multidict==6.4.3
packaging==24.2
pillow==11.2.1
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
PyJWT==2.10.1
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
qrcode==8.1
redis==5.2.1
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
twilio==9.5.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
yarl==1.19.0
